Introduce parse_json helper function
We need to relax the assumption that a column always has mixed NaN and well
formed JSON values. This only works when we use parse_json in KQL query.
When parse_json is not present in the KQL query the resulting missing values
are not NaN but empty strings.
For this reason, we introduce a parse_json function in the code that treats
empty strings correctly.